fixed the file structure, linked learn more to home page, created a home page
