+++ src/app/home/components/HomePage.tsx
@@ -1,0 +1,13 @@
+"use client";
+
+import React from 'react';
+
+const HomePage = () => {
+  return (
+    <div>
+      <h1>Welcome to the Homepage!</h1>
+    </div>
+  );
+};
+
+export default HomePage;
