+++ src/app/home/page.tsx
@@ -1,0 +1,6 @@
+
+import HomePageComponent from './components/HomePage'; 
+
+export default function HomePageRoute() { 
+  return <HomePageComponent />; 
+}
